refactor(database): apply CodeRabbit suggestions to databaseService
- Clear initializationPromise after successful initialization
- Fix log prefix duplication in migration onLog handler
- Use consistent DatabaseError types with proper error codes
- Extract migration context/adapter to separate module

Created src/database/adapters/expoSqliteAdapter.js to reduce
databaseService.js from 370+ lines to 218 lines, meeting
the 150-200 line guideline for better maintainability.
